requirements.txt
groq
langchain
langchain-community
langchain-text-splitters
chromadb
sentence-transformers
python-dotenv
pypdf